# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.AdminTeamResponse:
    properties:
      description:
        type: string
      display_order:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      logo_path:
        type: string
      name:
        type: string
    type: object
  github_com_locde0_sportudei-ukma_backend_internal_dto.AdminTeamsListResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminTeamResponse'
        type: array
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.BaseTeamResponse:
    properties:
      description:
        type: string
      display_order:
        type: integer
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
    type: object
...
  github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateTeamOrderRequest:
    properties:
      display_order:
        minimum: 0
        type: integer
    type: object
...
info:
  contact: {}
  description: sportudei api server
  title: sportudei api
  version: "1.0"
paths:
...
  /api/admin/teams:
    get:
      description: List all teams for admin view
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminTeamsListResponse'
      security:
      - BearerAuth: []
      summary: List admin teams
      tags:
      - admin-teams
    post:
      consumes:
      - multipart/form-data
      description: Create a new team with a logo
      parameters:
      - description: CreateTeamRequest JSON string
        in: formData
        name: payload
        required: true
        type: string
      - description: Team logo photo
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - admin-teams
  /api/admin/teams/{id}:
    delete:
      description: Delete an existing team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Delete team
      tags:
      - admin-teams
    get:
      description: Get full team details for admin by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.AdminTeamResponse'
      security:
      - BearerAuth: []
      summary: Get admin team
      tags:
      - admin-teams
    put:
      consumes:
      - multipart/form-data
      description: Update an existing team and its logo
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateTeamRequest JSON string
        in: formData
        name: payload
        required: true
        type: string
      - description: Team logo photo (optional)
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update team
      tags:
      - admin-teams
  /api/admin/teams/{id}/order:
    put:
      consumes:
      - application/json
      description: Update the display order of a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: New order
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.UpdateTeamOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Update team order
      tags:
      - admin-teams
...
  /api/teams/{id}:
    get:
      description: Get active team details for public by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_locde0_sportudei-ukma_backend_internal_dto.BaseTeamResponse'
      summary: Get public team
      tags:
      - public-teams
swagger: "2.0"